# services/ingestion-service/internal/driver/http/generated/server.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /admin/collect-subscriptions)
	AdminCollectSubscriptions(c *gin.Context, params AdminCollectSubscriptionsParams)

	// (POST /admin/collect-trending)
	AdminCollectTrending(c *gin.Context, params AdminCollectTrendingParams)

	// (POST /admin/schedule-snapshots)
	AdminScheduleSnapshots(c *gin.Context, params AdminScheduleSnapshotsParams)

	// (POST /admin/update-channels)
	AdminUpdateChannels(c *gin.Context, params AdminUpdateChannelsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// AdminCollectSubscriptions operation middleware
func (siw *ServerInterfaceWrapper) AdminCollectSubscriptions(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params AdminCollectSubscriptionsParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-CloudScheduler-JobName" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-JobName")]; found {
		var XCloudSchedulerJobName string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-JobName, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-JobName", valueList[0], &XCloudSchedulerJobName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-JobName: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerJobName = XCloudSchedulerJobName

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-JobName is required, but not found"), http.StatusBadRequest)
		return
	}

	// ------------- Required header parameter "X-CloudScheduler-ScheduleTime" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-ScheduleTime")]; found {
		var XCloudSchedulerScheduleTime string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-ScheduleTime, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-ScheduleTime", valueList[0], &XCloudSchedulerScheduleTime, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-ScheduleTime: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerScheduleTime = XCloudSchedulerScheduleTime

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-ScheduleTime is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminCollectSubscriptions(c, params)
}

// AdminCollectTrending operation middleware
func (siw *ServerInterfaceWrapper) AdminCollectTrending(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params AdminCollectTrendingParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-CloudScheduler-JobName" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-JobName")]; found {
		var XCloudSchedulerJobName string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-JobName, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-JobName", valueList[0], &XCloudSchedulerJobName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-JobName: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerJobName = XCloudSchedulerJobName

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-JobName is required, but not found"), http.StatusBadRequest)
		return
	}

	// ------------- Required header parameter "X-CloudScheduler-ScheduleTime" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-ScheduleTime")]; found {
		var XCloudSchedulerScheduleTime string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-ScheduleTime, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-ScheduleTime", valueList[0], &XCloudSchedulerScheduleTime, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-ScheduleTime: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerScheduleTime = XCloudSchedulerScheduleTime

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-ScheduleTime is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminCollectTrending(c, params)
}

// AdminScheduleSnapshots operation middleware
func (siw *ServerInterfaceWrapper) AdminScheduleSnapshots(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params AdminScheduleSnapshotsParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-CloudScheduler-JobName" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-JobName")]; found {
		var XCloudSchedulerJobName string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-JobName, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-JobName", valueList[0], &XCloudSchedulerJobName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-JobName: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerJobName = XCloudSchedulerJobName

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-JobName is required, but not found"), http.StatusBadRequest)
		return
	}

	// ------------- Required header parameter "X-CloudScheduler-ScheduleTime" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-ScheduleTime")]; found {
		var XCloudSchedulerScheduleTime string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-ScheduleTime, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-ScheduleTime", valueList[0], &XCloudSchedulerScheduleTime, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-ScheduleTime: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerScheduleTime = XCloudSchedulerScheduleTime

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-ScheduleTime is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminScheduleSnapshots(c, params)
}

// AdminUpdateChannels operation middleware
func (siw *ServerInterfaceWrapper) AdminUpdateChannels(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params AdminUpdateChannelsParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-CloudScheduler-JobName" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-JobName")]; found {
		var XCloudSchedulerJobName string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-JobName, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-JobName", valueList[0], &XCloudSchedulerJobName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-JobName: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerJobName = XCloudSchedulerJobName

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-JobName is required, but not found"), http.StatusBadRequest)
		return
	}

	// ------------- Required header parameter "X-CloudScheduler-ScheduleTime" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CloudScheduler-ScheduleTime")]; found {
		var XCloudSchedulerScheduleTime string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-CloudScheduler-ScheduleTime, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CloudScheduler-ScheduleTime", valueList[0], &XCloudSchedulerScheduleTime, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-CloudScheduler-ScheduleTime: %w", err), http.StatusBadRequest)
			return
		}

		params.XCloudSchedulerScheduleTime = XCloudSchedulerScheduleTime

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-CloudScheduler-ScheduleTime is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminUpdateChannels(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/admin/collect-subscriptions", wrapper.AdminCollectSubscriptions)
	router.POST(options.BaseURL+"/admin/collect-trending", wrapper.AdminCollectTrending)
	router.POST(options.BaseURL+"/admin/schedule-snapshots", wrapper.AdminScheduleSnapshots)
	router.POST(options.BaseURL+"/admin/update-channels", wrapper.AdminUpdateChannels)
}