# services/ingestion-service/internal/driver/http/generated/types.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package generated

// CollectSubscriptionsResponse defines model for CollectSubscriptionsResponse.
type CollectSubscriptionsResponse struct {
	ChannelsProcessed int32  `json:"channelsProcessed"`
	Duration          string `json:"duration"`
	VideosAdded       int32  `json:"videosAdded"`
	VideosProcessed   int32  `json:"videosProcessed"`
}

// CollectTrendingResponse defines model for CollectTrendingResponse.
type CollectTrendingResponse struct {
	Duration        string `json:"duration"`
	VideosAdded     int32  `json:"videosAdded"`
	VideosProcessed int32  `json:"videosProcessed"`
}

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// ScheduleSnapshotsResponse defines model for ScheduleSnapshotsResponse.
type ScheduleSnapshotsResponse struct {
	Duration        string `json:"duration"`
	TasksScheduled  int32  `json:"tasksScheduled"`
	VideosProcessed int32  `json:"videosProcessed"`
}

// UpdateChannelsResponse defines model for UpdateChannelsResponse.
type UpdateChannelsResponse struct {
	ChannelsProcessed int32  `json:"channelsProcessed"`
	ChannelsUpdated   int32  `json:"channelsUpdated"`
	Duration          string `json:"duration"`
}

// AdminCollectSubscriptionsParams defines parameters for AdminCollectSubscriptions.
type AdminCollectSubscriptionsParams struct {
	XCloudSchedulerJobName      string `json:"X-CloudScheduler-JobName"`
	XCloudSchedulerScheduleTime string `json:"X-CloudScheduler-ScheduleTime"`
}

// AdminCollectTrendingParams defines parameters for AdminCollectTrending.
type AdminCollectTrendingParams struct {
	XCloudSchedulerJobName      string `json:"X-CloudScheduler-JobName"`
	XCloudSchedulerScheduleTime string `json:"X-CloudScheduler-ScheduleTime"`
}

// AdminScheduleSnapshotsParams defines parameters for AdminScheduleSnapshots.
type AdminScheduleSnapshotsParams struct {
	XCloudSchedulerJobName      string `json:"X-CloudScheduler-JobName"`
	XCloudSchedulerScheduleTime string `json:"X-CloudScheduler-ScheduleTime"`
}

// AdminUpdateChannelsParams defines parameters for AdminUpdateChannels.
type AdminUpdateChannelsParams struct {
	XCloudSchedulerJobName      string `json:"X-CloudScheduler-JobName"`
	XCloudSchedulerScheduleTime string `json:"X-CloudScheduler-ScheduleTime"`
}